+++ Tests/TuistKitTests/Services/ScaffoldServiceTests.swift
@@ -10,6 +10,7 @@ import XCTest
 @testable import TuistCoreTesting
 @testable import TuistKit
 @testable import TuistLoaderTesting
+@testable import TuistPluginTesting
 @testable import TuistScaffoldTesting
 @testable import TuistSupportTesting
 
@@ -18,16 +19,19 @@ final class ScaffoldServiceTests: TuistUnitTestCase {
     var templateLoader: MockTemplateLoader!
     var templatesDirectoryLocator: MockTemplatesDirectoryLocator!
     var templateGenerator: MockTemplateGenerator!
+    var pluginService: MockPluginService!
 
     override func setUp() {
         super.setUp()
         templateLoader = MockTemplateLoader()
         templatesDirectoryLocator = MockTemplatesDirectoryLocator()
         templateGenerator = MockTemplateGenerator()
+        pluginService = MockPluginService()
         subject = ScaffoldService(
             templateLoader: templateLoader,
             templatesDirectoryLocator: templatesDirectoryLocator,
-            templateGenerator: templateGenerator
+            templateGenerator: templateGenerator,
+            pluginService: pluginService
         )
     }
 
@@ -60,6 +64,37 @@ final class ScaffoldServiceTests: TuistUnitTestCase {
         // When
         let options = try subject.loadTemplateOptions(
             templateName: "template",
+            path: nil
+        )
+
+        // Then
+        XCTAssertEqual(options.required, expectedOptions.required)
+        XCTAssertEqual(options.optional, expectedOptions.optional)
+    }
+
+    func test_load_template_plugin_options() throws {
+        // Given
+        let templatePath = try temporaryPath().appending(component: "PluginTemplate")
+
+        pluginService.loadPluginsStub = { _ in
+            Plugins.test(templatePaths: [templatePath])
+        }
+
+        templateLoader.loadTemplateStub = { _ in
+            Template(
+                description: "test",
+                attributes: [
+                    .required("required"),
+                    .optional("optional", default: ""),
+                ]
+            )
+        }
+
+        let expectedOptions: (required: [String], optional: [String]) = (required: ["required"], optional: ["optional"])
+
+        // When
+        let options = try subject.loadTemplateOptions(
+            templateName: "PluginTemplate",
             path: nil
         )